custom unit for image calibration + particles diameter measures

+++ src/main/java/fr/sorbonne_universite/ldc/model/workers/MeasuresProcessingWorker.java
@@ -28,6 +28,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     private boolean excludeOnEdgesEnabled;
     private double circularityThreshold;
     private boolean showAreaEnabled;
+    private boolean showDiameterEnabled;
     private boolean showMeanEnabled;
     private boolean showMedianEnabled;
     private boolean showIntegratedDensityEnabled;
@@ -38,18 +39,19 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
      * Creates a {@code MeasuresProcessingWorker}.
      * @param isCalibrated					Boolean that tell if the image there is a calibration given for the image
      * @param Calibration					Given calibration.
-     * @param minSize 						Minimum particle size (μm² IsCalibrated is true, otherwise px²).
-     * @param maxSize 						Maximum particle size (μm² IsCalibrated is true, otherwise px²).
+     * @param minSize 						Minimum particle size (unit² if IsCalibrated is true, otherwise px²).
+     * @param maxSize 						Maximum particle size (unit² if IsCalibrated is true, otherwise px²).
      * @param minCircularity 				Minimum particle circularity.
      * @param maxCircularity 				Maximum particle circularity.
      * @param excludeOnEdgesEnabled 		Particle Analyzer option.
      * @param circularityThreshold 			Threshold on particle's circularity to define if they are isolated. 
-     * @param showAreaEnabled 				True if the 'Area' column must be shown in the results.
-     * @param showMedianEnabled 			True if the 'Median' column must be shown in the results.
-     * @param showMeanEnabled 				True if the 'Mean' column must be shown in the results.
-     * @param showIntegratedDensityEnabled 	True if the 'IntegratedDensity' column must be shown in the results.
-     * @param showCircularityEnabled 		True if the 'Circularity' column is shown must be the results.
-     * @param img The current image to consider.
+     * @param showAreaEnabled 				If True then the 'Area' column must be shown in the results.
+     * @param showDiameterEnabled			If True then the 'Diameter' column must be shown in the results.
+     * @param showMedianEnabled 			If True then the 'Median' column must be shown in the results.
+     * @param showMeanEnabled 				If True then the 'Mean' column must be shown in the results.
+     * @param showIntegratedDensityEnabled 	If True then the 'IntegratedDensity' column must be shown in the results.
+     * @param showCircularityEnabled 		If True then the 'Circularity' column is shown must be the results.
+     * @param img 							The current image to consider.
      */
     public MeasuresProcessingWorker(
     		boolean isCalibrated,
@@ -61,6 +63,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     		boolean excludeOnEdgesEnabled,
     		double circularityThreshold,
     		boolean showAreaEnabled,
+    		boolean showDiameterEnabled,
     		boolean showMedianEnabled,
     		boolean showMeanEnabled,
     		boolean showIntegratedDensityEnabled,
@@ -75,6 +78,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     	this.excludeOnEdgesEnabled = excludeOnEdgesEnabled;
     	this.circularityThreshold = circularityThreshold;
     	this.showAreaEnabled = showAreaEnabled;
+    	this.showDiameterEnabled = showDiameterEnabled;
     	this.showMedianEnabled = showMedianEnabled;
     	this.showMeanEnabled = showMeanEnabled;
     	this.showIntegratedDensityEnabled = showIntegratedDensityEnabled;
@@ -92,6 +96,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     	if (showMeanEnabled) measurements += Measurements.MEAN;
     	if (showMedianEnabled) measurements += Measurements.MEDIAN;
     	if (showIntegratedDensityEnabled) measurements += Measurements.INTEGRATED_DENSITY;
+    	if (showDiameterEnabled) measurements += Measurements.ELLIPSE;
     	
     	// measure circularity even if the show circularity is disabled, to tell if a particle is isolated
     	measurements += Measurements.CIRCULARITY;
@@ -104,13 +109,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     	
     	// set options for Particles Analyzer
     	int options = 0;
-    	// options += ParticleAnalyzer.ELLIPSE;  // show overlay of detected particles
-    	// options += ParticleAnalyzer.OVERLAY; // crash the program
-    	// options += ParticleAnalyzer.SHOW_OUTLINES; // show outlines of every particles in each images of the stack
-    	// options += ParticleAnalyzer.DISPLAY_SUMMARY; // show statistics
-    	if (excludeOnEdgesEnabled) {
-    		options += ParticleAnalyzer.EXCLUDE_EDGE_PARTICLES;
-    	}
+    	if (excludeOnEdgesEnabled) options += ParticleAnalyzer.EXCLUDE_EDGE_PARTICLES;
 
     	// get current image
     	if (img == null) {
@@ -122,7 +121,7 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     	double pxMaxSize = maxSize;
     	
     	if (isCalibrated && calibration != null && calibration.scaled()) {
-    		// convert the µm² in px²
+    		// convert the unit² in px²
     		double pixelArea = calibration.pixelWidth * calibration.pixelHeight;
     		pxMinSize = minSize / pixelArea;
     		if (maxSize != Double.MAX_VALUE) {
@@ -189,19 +188,46 @@ public class MeasuresProcessingWorker extends SwingWorker<Void, Void>{
     	    		rt.setValue("is_isolated", row, isIsolated);
     	    	}
         	}
+        	
+    		// get the unit of the calibration
+    		String unit = "px";
+    		if (isCalibrated && calibration != null) unit = calibration.getUnit();
+    		
+    		// add the unit in the area column
+    		if (showAreaEnabled && rt.columnExists("Area")) {
+    			rt.renameColumn("Area", "Area(" + unit + "²)");
+    		}
+        	
+        	// compute the diameters
+        	if (showDiameterEnabled && rt.columnExists("Major") && rt.columnExists("Minor")) {
+        		int rowCount = rt.getCounter();
+        	    
+        		// get the values of big and small axes of the particles
+        		double[] majorAxis = rt.getColumnAsDoubles(rt.getColumnIndex("Major"));
+        		double[] minorAxis = rt.getColumnAsDoubles(rt.getColumnIndex("Minor"));
+        		
+        		// create the Diameter column with the correct unit
+        		int diameterIndex = rt.getFreeColumn("Diameter(" + unit + ")");
+        	    
+        		// compute and add the diameter for each particle
+        		for (int i = 0; i < rowCount; i++) {
+        			double diameter = (majorAxis[i] + minorAxis[i]) / 2.0;	// average of the two
+        			rt.setValue(diameterIndex, i, diameter);
+        		}
+        	}
 
         	// cleaning the results table : remove unused columns
         	if (rt.columnExists("BX")) rt.deleteColumn("BX");
             if (rt.columnExists("BY")) rt.deleteColumn("BY");
             if (rt.columnExists("Width")) rt.deleteColumn("Width");
             if (rt.columnExists("Height")) rt.deleteColumn("Height");
-            if (!showCircularityEnabled){
-            	rt.deleteColumn("Circ."); // remove if the circularity parameter isn't activated
-            	rt.deleteColumn("AR");
-            	rt.deleteColumn("Round");
-            	rt.deleteColumn("Solidity");
-            }
-        	
+            if (rt.columnExists("Major")) rt.deleteColumn("Major");
+            if (rt.columnExists("Minor")) rt.deleteColumn("Minor");
+            if (rt.columnExists("Angle")) rt.deleteColumn("Angle");
+            if (rt.columnExists("AR")) rt.deleteColumn("AR");
+            if (rt.columnExists("Round")) rt.deleteColumn("Round");
+            if (rt.columnExists("Solidity")) rt.deleteColumn("Solidity");
+            if (!showCircularityEnabled) rt.deleteColumn("Circ."); // remove if the circularity parameter isn't activated
             
     		// close the ROI manager window that appear with the ParticlesAnalyzer WIP
         	RoiManager rm = RoiManager.getInstance();
